refactor: moved user generatino script to place where other scripts are and reused DataGenerator method there.
